Move MyAdaTabs to /constants for use in /device
