+++ src/app/components/elements/sidebar/MyAdaSidebar.tsx
@@ -3,76 +3,9 @@ import { ContentSidebar, ContentSidebarProps } from "../layout/SidebarLayout";
 import { StyledTabPicker } from "../inputs/StyledTabPicker";
 import classNames from "classnames";
 import { selectors, sidebarSlice, useAppDispatch, useAppSelector } from "../../../state";
-import { above, below, isStudent, isTeacherOrAbove, isTutorOrAbove, useDeviceSize, useUserNotifications } from "../../../services";
+import { above, below, isStudent, isTeacherOrAbove, isTutorOrAbove, MyAdaTabs, useDeviceSize, useUserNotifications } from "../../../services";
 import { Spacer } from "../Spacer";
 import { useLocation } from "react-router";
-
-interface MyAdaTab {
-    title: string;
-    url: string;
-    icon: string;
-    user: "STUDENT" | "TUTOR" | "TEACHER" | "ALL"; // Which user roles can see this tab – n.b. teacher means teacherOrAbove
-}
-
-export const MyAdaTabs: Record<string, MyAdaTab> = {
-    overview: {
-        title: "Overview",
-        url: "/dashboard",
-        icon: "icon-home",
-        user: "ALL"
-    },
-    groups: {
-        title: "Groups",
-        url: "/groups",
-        icon: "icon-group",
-        user: "TUTOR"
-    },
-    setQuizzes: {
-        title: "Quizzes",
-        url: "/quizzes/set",
-        icon: "icon-file",
-        user: "TUTOR"
-    },
-    setTests: {
-        title: "Tests",
-        url: "/set_tests",
-        icon: "icon-school",
-        user: "TEACHER"
-    },
-    markbook: {
-        title: "Markbook",
-        url: "/my_markbook",
-        icon: "icon-done-all",
-        user: "TUTOR"
-    },
-
-    assignedToMe: {
-        title: "Assigned to me",
-        url: "/assignments",
-        icon: "icon-person-check",
-        user: "ALL"
-    },
-
-    myTests: {
-        title: "Tests",
-        url: "/tests",
-        icon: "icon-school",
-        user: "STUDENT"
-    },
-    progress: {
-        title: "Progress",
-        url: "/progress",
-        icon: "icon-done-all",
-        user: "STUDENT"
-    },
-
-    account: {
-        title: "Account",
-        url: "/account",
-        icon: "icon-cog",
-        user: "ALL"
-    }
-};
 
 interface AdaSidebarCollapserProps extends React.HTMLAttributes<HTMLButtonElement> {
     collapsed: boolean;
